Refactor file structure in editor folder
